chore(ZNTA-1865): extract TMMatchPercentageDropdown into own component
